pyproject.toml
[tool.poetry]
name = "acli"
version = "0.1.0"
description = "A small CLI to help me easily log my Aggietime hours."
authors = ["Jaren Glenn <jarenglenn@gmail.com>"]
packages = [
    {include = "acli", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.9"
arc-cli = "^2.3.0"
beautifulsoup4 = "^4.9.3"
keyring = "^23.0.1"
requests = "^2.25.1"
html5lib = "^1.1"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
mypy = "^0.901"
mypy-extensions = "^0.4.3"

[tool.poetry.scripts]
acli = "acli.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
